refactored reactive values from store into the photostitching feature folder. It should be used only for that feature
